Added fiex

+++ openai_ros/src/openai_ros/task_envs/turtlebot3/turtlebot3_world.py
@@ -39,11 +39,11 @@ class TurtleBot3WorldEnv(turtlebot3_env.TurtleBot3Env):
         # Only variable needed to be set here
         number_actions = rospy.get_param('/turtlebot3/n_actions')
         self.action_space = spaces.Discrete(number_actions)
-        
+
         # We set the reward range, which is not compulsory but here we do it.
         self.reward_range = (-numpy.inf, numpy.inf)
-        
-        
+
+
         #number_observations = rospy.get_param('/turtlebot3/n_observations')
         """
         We set the Observation space for the 6 observations
@@ -56,34 +56,34 @@ class TurtleBot3WorldEnv(turtlebot3_env.TurtleBot3Env):
             round(yaw, 1),
         ]
         """
-        
+
         # Actions and Observations
         self.linear_forward_speed = rospy.get_param('/turtlebot3/linear_forward_speed')
         self.linear_turn_speed = rospy.get_param('/turtlebot3/linear_turn_speed')
         self.angular_speed = rospy.get_param('/turtlebot3/angular_speed')
         self.init_linear_forward_speed = rospy.get_param('/turtlebot3/init_linear_forward_speed')
         self.init_linear_turn_speed = rospy.get_param('/turtlebot3/init_linear_turn_speed')
-        
+
         self.new_ranges = rospy.get_param('/turtlebot3/new_ranges')
         self.min_range = rospy.get_param('/turtlebot3/min_range')
         self.max_laser_value = rospy.get_param('/turtlebot3/max_laser_value')
         self.min_laser_value = rospy.get_param('/turtlebot3/min_laser_value')
         self.max_linear_aceleration = rospy.get_param('/turtlebot3/max_linear_aceleration')
-        
-        
+
+
         # We create two arrays based on the binary values that will be assigned
         # In the discretization method.
         laser_scan = self.get_laser_scan()
-        num_laser_readings = len(laser_scan.ranges)/self.new_ranges
+        num_laser_readings = int(len(laser_scan.ranges)/self.new_ranges)
         high = numpy.full((num_laser_readings), self.max_laser_value)
         low = numpy.full((num_laser_readings), self.min_laser_value)
-        
+
         # We only use two integers
         self.observation_space = spaces.Box(low, high)
-        
+
         rospy.logdebug("ACTION SPACES TYPE===>"+str(self.action_space))
         rospy.logdebug("OBSERVATION SPACES TYPE===>"+str(self.observation_space))
-        
+
         # Rewards
         self.forwards_reward = rospy.get_param("/turtlebot3/forwards_reward")
         self.turn_reward = rospy.get_param("/turtlebot3/turn_reward")
@@ -121,7 +121,7 @@ class TurtleBot3WorldEnv(turtlebot3_env.TurtleBot3Env):
         based on the action number given.
         :param action: The action integer that set s what movement to do next.
         """
-        
+
         rospy.logdebug("Start Set Action ==>"+str(action))
         # We convert the actions to speed movements to send to the parent class CubeSingleDiskEnv
         if action == 0: #FORWARD
@@ -136,10 +136,10 @@ class TurtleBot3WorldEnv(turtlebot3_env.TurtleBot3Env):
             linear_speed = self.linear_turn_speed
             angular_speed = -1*self.angular_speed
             self.last_action = "TURN_RIGHT"
-        
+
         # We tell TurtleBot2 the linear and angular speed to set to execute
         self.move_base(linear_speed, angular_speed, epsilon=0.05, update_rate=10)
-        
+
         rospy.logdebug("END Set Action ==>"+str(action))
 
     def _get_obs(self):
@@ -152,7 +152,7 @@ class TurtleBot3WorldEnv(turtlebot3_env.TurtleBot3Env):
         rospy.logdebug("Start Get Observation ==>")
         # We get the laser scan data
         laser_scan = self.get_laser_scan()
-        
+
         discretized_observations = self.discretize_scan_observation(    laser_scan,
                                                                         self.new_ranges
                                                                         )
@@ -160,15 +160,15 @@ class TurtleBot3WorldEnv(turtlebot3_env.TurtleBot3Env):
         rospy.logdebug("Observations==>"+str(discretized_observations))
         rospy.logdebug("END Get Observation ==>")
         return discretized_observations
-        
+
 
     def _is_done(self, observations):
-        
+
         if self._episode_done:
             rospy.logerr("TurtleBot2 is Too Close to wall==>")
         else:
             rospy.logwarn("TurtleBot2 is NOT close to a wall ==>")
-            
+
         # Now we check if it has crashed based on the imu
         imu_data = self.get_imu()
         linear_acceleration_magnitude = self.get_vector_magnitude(imu_data.linear_acceleration)
@@ -177,7 +177,7 @@ class TurtleBot3WorldEnv(turtlebot3_env.TurtleBot3Env):
             self._episode_done = True
         else:
             rospy.logerr("DIDNT crash TurtleBot2 ==>"+str(linear_acceleration_magnitude)+">"+str(self.max_linear_aceleration))
-        
+
 
         return self._episode_done
 
@@ -197,26 +197,26 @@ class TurtleBot3WorldEnv(turtlebot3_env.TurtleBot3Env):
         rospy.logdebug("Cumulated_reward=" + str(self.cumulated_reward))
         self.cumulated_steps += 1
         rospy.logdebug("Cumulated_steps=" + str(self.cumulated_steps))
-        
+
         return reward
 
 
     # Internal TaskEnv Methods
-    
+
     def discretize_scan_observation(self,data,new_ranges):
         """
         Discards all the laser readings that are not multiple in index of new_ranges
         value.
         """
         self._episode_done = False
-        
+
         discretized_ranges = []
         mod = len(data.ranges)/new_ranges
-        
+
         rospy.logdebug("data=" + str(data))
         rospy.logdebug("new_ranges=" + str(new_ranges))
         rospy.logdebug("mod=" + str(mod))
-        
+
         for i, item in enumerate(data.ranges):
             if (i%mod==0):
                 if item == float ('Inf') or numpy.isinf(item):
@@ -225,21 +225,21 @@ class TurtleBot3WorldEnv(turtlebot3_env.TurtleBot3Env):
                     discretized_ranges.append(self.min_laser_value)
                 else:
                     discretized_ranges.append(int(item))
-                    
+
                 if (self.min_range > item > 0):
                     rospy.logerr("done Validation >>> item=" + str(item)+"< "+str(self.min_range))
                     self._episode_done = True
                 else:
                     rospy.logdebug("NOT done Validation >>> item=" + str(item)+"< "+str(self.min_range))
-                    
+
 
         return discretized_ranges
-        
-        
+
+
     def get_vector_magnitude(self, vector):
         """
         It calculated the magnitude of the Vector3 given.
-        This is usefull for reading imu accelerations and knowing if there has been 
+        This is usefull for reading imu accelerations and knowing if there has been
         a crash
         :return:
         """
